Set up opening and closing an mlx window
To compile with the mlx library and not run into issues where some
functions couldn't be found, I had to explicitly set the compiler to
clang in the makefile. But this also meant some lines got flagged as an
error by clang. It was when a float was checked against an INT_MIN or
INT_MAX causing. I've temporarily typecast them to a float in order to
test the program.

To check if creating and destroying the window leaves no leaks, I tested
with valgrind which my version can't handle the clang dwarf-5 version.
So I've put the flag to dwarf-4 during compilation. After checking, no
leaks were found.

+++ src/parse_coordinates.c
@@ -6,7 +6,7 @@
 /*   By: pclaus <pclaus@student.s19.be>             +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/08/05 12:27:26 by pclaus            #+#    #+#             */
-/*   Updated: 2024/08/09 10:20:26 by pclaus           ###   ########.fr       */
+/*   Updated: 2024/08/17 21:52:38 by efret            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 static int	check_limits(float coordinate_value, char **coordinates_split)
 {
-	if (coordinate_value < INT_MIN || coordinate_value > INT_MAX)
+	if (coordinate_value < (float)INT_MIN || coordinate_value > (float)INT_MAX)
 	{
 		free_split(coordinates_split);
 		return (1);
